+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/PhysicalResourceDetails.test.js
@@ -32,6 +32,10 @@ jest.mock('../../../hooks', () => ({
       value: '',
     }],
   }]],
+  useDisabledOrderFields: () => ({
+    dismissCreateInventory: false,
+    dismissPhysicalDetails: false,
+  }),
 }));
 
 const setReferenceTablesMock = jest.fn();