+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugs fixed:
 * Order field mapping profile: Fix the View UI for Fund value % and Currency (UIDATIMP-1404)
 * Order field mapping profile: Fix the View UI for Discount Currency (UIDATIMP-1405)
+* Order field mapping profile: Inactivate Locations and Material Types fields when Order status is Open (UIDATIMP-1417)
 
 ## [6.0.2](https://github.com/folio-org/ui-data-import/tree/v6.0.2) (2023-03-24)